requirements.txt
fastapi
uvicorn
pdfplumber
pydantic>=2.0
python-dateutil
typer[all]
rich